code format adjust

+++ backend/arm32/PlatformArm32.h
@@ -52,7 +52,7 @@ public:
     static const int maxRegNum = 16;
 
     /// @brief 可使用的通用寄存器的个数r0-r10
-    static const unsigned long regNum = 11;
+    static const int regNum = 11;
 
     /// @brief 寄存器的名字，r0-r15
     static const std::string regName[maxRegNum];
